move get_student_attendance.py up a dir
